# More robust handling of maximum texture size for non-color data, slight perf improvements for large point clouds