--- a/crates/re_renderer/src/allocator/data_texture_source.rs
+++ b/crates/re_renderer/src/allocator/data_texture_source.rs
@@ -7,6 +7,22 @@ use crate::{
 
 use super::{CpuWriteGpuReadBuffer, CpuWriteGpuReadError};
 
+#[derive(thiserror::Error, Debug, PartialEq, Eq)]
+pub enum DataTextureSourceWriteError {
+    #[error(
+        "Reached maximum number of elements for a data texture of {max_num_elements} elements.
+ Tried to add {num_elements_attempted_to_add} elements, but only added {num_elements_actually_added}."
+    )]
+    ReachedMaximumNumberOfElements {
+        max_num_elements: usize,
+        num_elements_attempted_to_add: usize,
+        num_elements_actually_added: usize,
+    },
+
+    #[error(transparent)]
+    CpuWriteGpuReadError(#[from] crate::CpuWriteGpuReadError),
+}
+
 /// Utility for writing data to a dynamically sized "data textures".
 ///
 /// For WebGL compatibility we sometimes have to use textures instead of buffers.
@@ -62,31 +78,6 @@ impl<'a, T: Pod + Send + Sync> DataTextureSource<'a, T> {
             buffers: Vec::new(),
             active_buffer_index: 0,
         }
-    }
-
-    /// Maximum width for data textures.
-    #[inline]
-    pub fn max_data_texture_width(device_limits: &wgpu::Limits) -> usize {
-        // We limit the data texture width to 16384 or whatever smaller value is supported but the device.
-        //
-        // If we make buffers always a multiple of this width, we can do all buffer copies in a single copy!
-        //
-        // But wait! If we're using this as the minimum buffer size, isn't that too big?
-        // 16384 * float4 (worst case) = 256KiB.
-        // Keep in mind that weaker hardware will have 8192 max width.
-        // Also note, that many of our textures use 4 & 8 byte formats.
-        // So while this is still a considerable amount of memory when used for very small data textures
-        // it's not as bad as it seems.
-        // Given how much it simplifies to keep buffers a multiple of the texture width,
-        // this seems to be a reasonable trade-off.
-        (device_limits.max_texture_dimension_2d as usize).min(16384)
-    }
-
-    /// Maximum number of elements that can be written to a single data texture.
-    pub fn max_num_elements(device_limits: &wgpu::Limits) -> usize {
-        let max_width = Self::max_data_texture_width(device_limits);
-        let max_height = device_limits.max_texture_dimension_2d as usize;
-        max_width * max_height
     }
 
     /// Whether no elements have been written at all.
@@ -155,51 +146,79 @@ impl<'a, T: Pod + Send + Sync> DataTextureSource<'a, T> {
 
     /// Ensures that there's space internally for at least `num_elements` more elements.
     ///
+    /// Returns the number of elements that are currently reserved.
+    /// This value is *smaller* than the requested number of elements if the maximum number of
+    /// elements that can be stored is reached, see [`max_num_elements_per_data_texture`].
+    ///
     /// Creating new buffers is a relatively expensive operation, so it's best to
     /// reserve gratuitously and often. Ideally, you know exactly how many elements you're going to write and reserve
     /// accordingly at the start.
     ///
     /// If there's no more space, a new buffer is allocated such that:
-    /// * have a total capacity for at least as many elements as requested
+    /// * have a total capacity for at least as many elements as requested, clamping total size to [`max_num_elements_per_data_texture`]
     /// * be at least double the size of the last buffer
     /// * keep it easy to copy to textures by always being a multiple of the maximum row size we use for data textures
     ///      -> this massively simplifies the buffer->texture copy logic!
-    pub fn reserve(&mut self, num_elements: usize) -> Result<(), CpuWriteGpuReadError> {
+    pub fn reserve(&mut self, num_elements: usize) -> Result<usize, CpuWriteGpuReadError> {
         let remaining_capacity = self.remaining_capacity();
         if remaining_capacity >= num_elements {
-            return Ok(());
+            return Ok(remaining_capacity);
         }
+
+        let max_texture_dimension_2d = self.ctx.device.limits().max_texture_dimension_2d;
 
         let last_buffer_size = self.buffers.last().map_or(0, |b| b.capacity());
         let new_buffer_size = (num_elements - remaining_capacity)
             .max(last_buffer_size * 2)
-            .next_multiple_of(Self::max_data_texture_width(&self.ctx.device.limits()));
+            .next_multiple_of(max_data_texture_width(max_texture_dimension_2d) as usize)
+            .min(max_num_elements_per_data_texture(max_texture_dimension_2d) - self.len());
 
-        self.buffers
-            .push(self.ctx.cpu_write_gpu_read_belt.lock().allocate(
-                &self.ctx.device,
-                &self.ctx.gpu_resources.buffers,
-                new_buffer_size,
-            )?);
+        if new_buffer_size > 0 {
+            self.buffers
+                .push(self.ctx.cpu_write_gpu_read_belt.lock().allocate(
+                    &self.ctx.device,
+                    &self.ctx.gpu_resources.buffers,
+                    new_buffer_size,
+                )?);
+        }
 
-        Ok(())
+        Ok(remaining_capacity + new_buffer_size)
+    }
+
+    fn error_on_clamped_write(
+        &self,
+        num_elements_attempted_to_add: usize,
+        num_elements_actually_added: usize,
+    ) -> Result<(), DataTextureSourceWriteError> {
+        if num_elements_actually_added < num_elements_attempted_to_add {
+            Err(
+                DataTextureSourceWriteError::ReachedMaximumNumberOfElements {
+                    max_num_elements: max_num_elements_per_data_texture(
+                        self.ctx.device.limits().max_texture_dimension_2d,
+                    ),
+                    num_elements_attempted_to_add,
+                    num_elements_actually_added,
+                },
+            )
+        } else {
+            Ok(())
+        }
     }
 
     /// Pushes a slice of elements into the data texture builder.
-    #[inline]
-    pub fn extend_from_slice(&mut self, elements: &[T]) -> Result<(), CpuWriteGpuReadError> {
+    pub fn extend_from_slice(&mut self, elements: &[T]) -> Result<(), DataTextureSourceWriteError> {
         if elements.is_empty() {
             return Ok(());
         }
 
         re_tracing::profile_function!();
 
-        self.reserve(elements.len())?;
+        let num_elements_available = self.reserve(elements.len())?;
+        let num_elements_actually_added = num_elements_available.min(elements.len());
 
-        let mut remaining_elements = elements;
-
+        let mut remaining_elements = &elements[..num_elements_actually_added];
         loop {
-            let result =
+            let write_result =
                 self.buffers[self.active_buffer_index].extend_from_slice(remaining_elements);
 
             // `extend_from_slice` is documented to write as many elements as possible, so we can just continue with the next buffer,
@@ -207,32 +226,36 @@ impl<'a, T: Pod + Send + Sync> DataTextureSource<'a, T> {
             if let Err(CpuWriteGpuReadError::BufferFull {
                 num_elements_actually_added,
                 ..
-            }) = result
+            }) = write_result
             {
                 remaining_elements = &remaining_elements[num_elements_actually_added..];
                 self.active_buffer_index += 1; // Due to the prior `reserve` call we know that there's more buffers!
             } else {
+                write_result?;
                 self.ensure_active_buffer_invariant_after_adding_elements();
-                return result;
+                return self.error_on_clamped_write(elements.len(), num_elements_actually_added);
             }
         }
     }
 
     /// Fills the buffer with n instances of an element.
-    #[inline]
-    pub fn add_n(&mut self, element: T, num_elements: usize) -> Result<(), CpuWriteGpuReadError> {
+    pub fn add_n(
+        &mut self,
+        element: T,
+        num_elements: usize,
+    ) -> Result<(), DataTextureSourceWriteError> {
         if num_elements == 0 {
             return Ok(());
         }
 
         re_tracing::profile_function!();
 
-        self.reserve(num_elements)?;
+        let num_elements_available = self.reserve(num_elements)?;
+        let num_elements_actually_added = num_elements_available.min(num_elements);
 
-        let mut num_elements_remaining = num_elements;
-
+        let mut num_elements_remaining = num_elements_actually_added;
         loop {
-            let result =
+            let write_result =
                 self.buffers[self.active_buffer_index].add_n(element, num_elements_remaining);
 
             // `fill_n` is documented to write as many elements as possible, so we can just continue with the next buffer,
@@ -240,29 +263,74 @@ impl<'a, T: Pod + Send + Sync> DataTextureSource<'a, T> {
             if let Err(CpuWriteGpuReadError::BufferFull {
                 num_elements_actually_added,
                 ..
-            }) = result
+            }) = write_result
             {
                 num_elements_remaining -= num_elements_actually_added;
                 self.active_buffer_index += 1; // Due to the prior `reserve` call we know that there's more buffers!
             } else {
+                write_result?;
                 self.ensure_active_buffer_invariant_after_adding_elements();
-                return result;
+                return self.error_on_clamped_write(num_elements, num_elements_actually_added);
             }
         }
     }
 
     #[allow(unused)] // TODO(andreas): soon!
     #[inline]
-    pub fn push(&mut self, element: T) -> Result<(), CpuWriteGpuReadError> {
-        self.reserve(1)?;
+    pub fn push(&mut self, element: T) -> Result<(), DataTextureSourceWriteError> {
+        if self.reserve(1)? < 1 {
+            return self.error_on_clamped_write(1, 0);
+        }
+
         self.buffers[self.active_buffer_index].push(element);
         self.ensure_active_buffer_invariant_after_adding_elements();
 
         Ok(())
     }
 
+    fn data_texture_size(&self, max_texture_dimension_2d: u32) -> wgpu::Extent3d {
+        let texel_size_in_bytes = std::mem::size_of::<T>() as u32;
+        let num_texels = self.len();
+
+        debug_assert!(num_texels < max_num_elements_per_data_texture(max_texture_dimension_2d));
+
+        // Our data textures are usually accessed in a linear fashion, so ideally we'd be using a 1D texture.
+        // However, 1D textures are very limited in size on many platforms, we have to use 2D textures instead.
+        // 2D textures perform a lot better when their dimensions are powers of two, so we'll strictly stick to that even
+        // when it seems to cause memory overhead.
+
+        // We fill row by row. With the power-of-two requirement, this is the optimal strategy:
+        // if there were a texture with less padding that uses half the width,
+        // then we'd need to increase the height. We can't increase without doubling it, thus creating a texture
+        // with the exact same mount of padding as before.
+        let max_data_texture_width = max_data_texture_width(max_texture_dimension_2d);
+        let width = if num_texels < max_data_texture_width as usize {
+            num_texels
+                .next_power_of_two()
+                // For too few number of written texels, or too small texels we might need to increase the size to stay
+                // above a row **byte** size of `wgpu::COPY_BYTES_PER_ROW_ALIGNMENT`.
+                // Note that this implies that for very large texels, we need less wide textures to stay above this limit.
+                // (width is in number of texels, but alignment cares about bytes!)
+                .next_multiple_of(
+                    (wgpu::COPY_BYTES_PER_ROW_ALIGNMENT / texel_size_in_bytes) as usize,
+                ) as u32
+        } else {
+            max_data_texture_width
+        };
+
+        let height = num_texels.div_ceil(width as usize);
+        debug_assert!(height <= max_texture_dimension_2d as usize); // Texel count should have been clamped accordingly already!
+
+        wgpu::Extent3d {
+            width,
+            height: height as u32,
+            depth_or_array_layers: 1,
+        }
+    }
+
     /// Schedules copies of all previous writes to this `DataTextureSource` to a `GpuTexture`.
     ///
+    /// The format *has* to be uncompressed, not a depth/stencil format and have the exact same block size of the size of type `T`.
     /// The resulting `GpuTexture` is ready to be bound as a data texture in a shader.
     pub fn finish(
         self,
@@ -271,20 +339,31 @@ impl<'a, T: Pod + Send + Sync> DataTextureSource<'a, T> {
     ) -> Result<GpuTexture, CpuWriteGpuReadError> {
         re_tracing::profile_function!();
 
-        let total_num_elements = self.len();
-
-        let texture_desc = data_texture_desc(
-            texture_label,
-            texture_format,
-            total_num_elements,
-            Self::max_data_texture_width(&self.ctx.device.limits()) as u32,
+        debug_assert!(!texture_format.has_depth_aspect());
+        debug_assert!(!texture_format.has_stencil_aspect());
+        debug_assert!(!texture_format.is_compressed());
+        debug_assert_eq!(
+            texture_format
+                .block_copy_size(None)
+                .expect("Depth/stencil formats are not supported as data textures"),
+            std::mem::size_of::<T>() as u32,
         );
-        let data_texture = self
-            .ctx
-            .gpu_resources
-            .textures
-            .alloc(&self.ctx.device, &texture_desc);
-        let texture_width = texture_desc.size.width as usize;
+
+        let texture_size =
+            self.data_texture_size(self.ctx.device.limits().max_texture_dimension_2d);
+        let data_texture = self.ctx.gpu_resources.textures.alloc(
+            &self.ctx.device,
+            &wgpu_resources::TextureDesc {
+                label: texture_label.into(),
+                size: texture_size,
+                mip_level_count: 1,
+                sample_count: 1,
+                dimension: wgpu::TextureDimension::D2,
+                format: texture_format,
+                usage: wgpu::TextureUsages::TEXTURE_BINDING | wgpu::TextureUsages::COPY_DST,
+            },
+        );
+        let texture_width = texture_size.width as usize;
 
         // Copy all buffers to the texture.
         let mut current_row = 0;
@@ -316,7 +395,7 @@ impl<'a, T: Pod + Send + Sync> DataTextureSource<'a, T> {
                 },
                 wgpu::Extent3d {
                     height: num_rows as u32,
-                    ..texture_desc.size
+                    ..texture_size
                 },
             )?;
 
@@ -327,6 +406,7 @@ impl<'a, T: Pod + Send + Sync> DataTextureSource<'a, T> {
     }
 }
 
+// TODO(andreas): This is now redundant to `DataTextureSource::data_texture_size`. Remove this once all line data has moved to `DataTextureSource`.
 /// Texture size for storing a given amount of data.
 ///
 /// For WebGL compatibility we sometimes have to use textures instead of buffers.
@@ -340,7 +420,6 @@ impl<'a, T: Pod + Send + Sync> DataTextureSource<'a, T> {
 /// row size in bytes is a multiple of `wgpu::COPY_BYTES_PER_ROW_ALIGNMENT`.
 /// This makes it a lot easier to copy data from a continuous buffer to the texture.
 /// If we wouldn't do that, we'd need to do a copy for each row in some cases.
-// TODO(andreas): everything should use `DataTextureSource` directly, then this function is no longer needed!
 pub fn data_texture_size(
     format: wgpu::TextureFormat,
     num_texels_written: usize,
@@ -411,4 +490,30 @@ pub fn data_texture_desc(
         format,
         usage: wgpu::TextureUsages::TEXTURE_BINDING | wgpu::TextureUsages::COPY_DST,
     }
+}
+
+/// Maximum width for data textures.
+#[inline]
+fn max_data_texture_width(max_texture_dimension_2d: u32) -> u32 {
+    // We limit the data texture width to 16384 or whatever smaller value is supported but the device.
+    //
+    // If we make buffers always a multiple of this width, we can do all buffer copies in a single copy!
+    //
+    // But wait! If we're using this as the minimum buffer size, isn't that too big?
+    // 16384 * float4 (worst case) = 256KiB.
+    // Keep in mind that weaker hardware will have 8192 max width.
+    // Also note, that many of our textures use 4 & 8 byte formats.
+    // So while this is still a considerable amount of memory when used for very small data textures
+    // it's not as bad as it seems.
+    // Given how much it simplifies to keep buffers a multiple of the texture width,
+    // this seems to be a reasonable trade-off.
+    (max_texture_dimension_2d).min(16384)
+}
+
+/// Maximum number of elements that can be written to a single data texture.
+#[inline]
+fn max_num_elements_per_data_texture(max_texture_dimension_2d: u32) -> usize {
+    let max_width = max_data_texture_width(max_texture_dimension_2d) as usize;
+    let max_height = max_texture_dimension_2d as usize;
+    max_width * max_height
 }