--- a/crates/re_renderer/src/allocator/data_texture_source.rs
+++ b/crates/re_renderer/src/allocator/data_texture_source.rs
@@ -84,7 +84,7 @@ impl<'a, T: Pod + Send + Sync> DataTextureSource<'a, T> {
     pub fn is_empty(&self) -> bool {
         self.buffers
             .first()
-            .map_or(false, |first_buffer| first_buffer.is_empty())
+            .map_or(true, |first_buffer| first_buffer.is_empty())
     }
 
     /// The number of elements written so far.