--- a/crates/re_renderer/src/allocator/mod.rs
+++ b/crates/re_renderer/src/allocator/mod.rs
@@ -11,7 +11,7 @@ mod uniform_buffer_fill;
 pub use cpu_write_gpu_read_belt::{
     CpuWriteGpuReadBelt, CpuWriteGpuReadBuffer, CpuWriteGpuReadError,
 };
-pub use data_texture_source::{data_texture_desc, data_texture_size, DataTextureSource};
+pub use data_texture_source::{data_texture_desc, DataTextureSource};
 pub use gpu_readback_belt::{
     GpuReadbackBelt, GpuReadbackBuffer, GpuReadbackError, GpuReadbackIdentifier,
 };