@Wumpf 2024-02-20
(tested on firefox webgl, chrome webgpu and ofc native)